# services/llm-api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenAI-compatible LLM API platform with enterprise authentication, conversation management, and streaming support.",
        "title": "Jan Server LLM API",
        "contact": {
            "name": "Jan Server Team",
            "url": "https://github.com/janhq/jan-server"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/admin/providers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all providers with their model counts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Provider API"
                ],
                "summary": "Get all providers",
                "responses": {
                    "200": {
                        "description": "List of providers with model counts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/modelresponses.ProviderWithModelCountResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve providers",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new provider and synchronizes its available models.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Provider API"
                ],
                "summary": "Register a provider",
                "parameters": [
                    {
                        "description": "Provider registration payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.AddProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered provider with synced models",
                        "schema": {
                            "$ref": "#/definitions/modelresponses.ProviderWithModelsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to register provider",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/providers/{provider_public_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing provider's configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Provider API"
                ],
                "summary": "Update a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider public ID",
                        "name": "provider_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Provider update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UpdateProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated provider",
                        "schema": {
                            "$ref": "#/definitions/modelresponses.ProviderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Provider not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update provider",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a model response for the given chat conversation. This is a standard chat completion API that supports both streaming and non-streaming modes without conversation persistence.\n\n**Streaming Mode (stream=true):**\n- Returns Server-Sent Events (SSE) with real-time streaming\n- Streams completion chunks directly from the inference model\n- Final event contains \"[DONE]\" marker\n\n**Non-Streaming Mode (stream=false or omitted):**\n- Returns single JSON response with complete completion\n- Standard OpenAI ChatCompletionResponse format\n\n**Storage Options:**\n- `store=true`: Persist the latest input message and assistant response to the active conversation\n- `store_reasoning=true`: Additionally persist reasoning content provided by the model\n- When `store` is omitted or false, the conversation remains read-only\n\n**Features:**\n- Supports all OpenAI ChatCompletionRequest parameters\n- Optional conversation context for conversation persistence\n- User authentication required\n- Direct inference model integration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "Chat Completions API"
                ],
                "summary": "Create a chat completion",
                "parameters": [
                    {
                        "description": "Chat completion request with streaming options and optional conversation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chatrequests.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful streaming response (when stream=true) - SSE format with data: {json} events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload, empty messages, or inference failure",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid authentication",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/models/providers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of available model providers that can be used for inference.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model API"
                ],
                "summary": "List model providers",
                "responses": {
                    "200": {
                        "description": "List of providers",
                        "schema": {
                            "$ref": "#/definitions/modelresponses.ProviderResponseList"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve providers",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "chatrequests.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "chat_template_kwargs": {
                    "description": "ChatTemplateKwargs provides a way to add non-standard parameters to the request body.\nAdditional kwargs to pass to the template renderer. Will be accessible by the chat template.\nSuch as think mode for qwen3. \"chat_template_kwargs\": {\"enable_thinking\": false}\nhttps://qwen.readthedocs.io/en/latest/deployment/vllm.html#thinking-non-thinking-modes",
                    "type": "object",
                    "additionalProperties": {}
                },
                "conversation": {
                    "description": "Conversation can be either a string (conversation ID) or a conversation object\nItems from this conversation are prepended to Messages for this response request.\nInput items and output items from this response are automatically added to this conversation after completion.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/chatrequests.ConversationReference"
                        }
                    ]
                },
                "frequency_penalty": {
                    "type": "number"
                },
                "function_call": {
                    "description": "Deprecated: use ToolChoice instead."
                },
                "functions": {
                    "description": "Deprecated: use Tools instead.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/openai.FunctionDefinition"
                    }
                },
                "guided_choice": {
                    "description": "GuidedChoice is a vLLM-specific extension that restricts the model's output\nto one of the predefined string choices provided in this field. This feature\nis used to constrain the model's responses to a controlled set of options,\nensuring predictable and consistent outputs in scenarios where specific\nchoices are required.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "logit_bias": {
                    "description": "LogitBias is must be a token id string (specified by their token ID in the tokenizer), not a word string.\nincorrect: `\"logit_bias\":{\"You\": 6}`, correct: `\"logit_bias\":{\"1639\": 6}`\nrefs: https://platform.openai.com/docs/api-reference/chat/create#chat/create-logit_bias",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "logprobs": {
                    "description": "LogProbs indicates whether to return log probabilities of the output tokens or not.\nIf true, returns the log probabilities of each output token returned in the content of message.\nThis option is currently not available on the gpt-4-vision-preview model.",
                    "type": "boolean"
                },
                "max_completion_tokens": {
                    "description": "MaxCompletionTokens An upper bound for the number of tokens that can be generated for a completion,\nincluding visible output tokens and reasoning tokens https://platform.openai.com/docs/guides/reasoning",
                    "type": "integer"
                },
                "max_tokens": {
                    "description": "MaxTokens The maximum number of tokens that can be generated in the chat completion.\nThis value can be used to control costs for text generated via API.\nDeprecated: use MaxCompletionTokens. Not compatible with o1-series models.\nrefs: https://platform.openai.com/docs/api-reference/chat/create#chat-create-max_tokens",
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/openai.ChatCompletionMessage"
                    }
                },
                "metadata": {
                    "description": "Metadata to store with the completion.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "model": {
                    "type": "string"
                },
                "n": {
                    "type": "integer"
                },
                "parallel_tool_calls": {
                    "description": "Disable the default behavior of parallel tool calls by setting it: false."
                },
                "prediction": {
                    "description": "Configuration for a predicted output.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/openai.Prediction"
                        }
                    ]
                },
                "presence_penalty": {
                    "type": "number"
                },
                "reasoning_effort": {
                    "description": "Controls effort on reasoning for reasoning models. It can be set to \"low\", \"medium\", or \"high\".",
                    "type": "string"
                },
                "response_format": {
                    "$ref": "#/definitions/openai.ChatCompletionResponseFormat"
                },
                "safety_identifier": {
                    "description": "A stable identifier used to help detect users of your application that may be violating OpenAI's usage policies.\nThe IDs should be a string that uniquely identifies each user.\nWe recommend hashing their username or email address, in order to avoid sending us any identifying information.\nhttps://platform.openai.com/docs/api-reference/chat/create#chat_create-safety_identifier",
                    "type": "string"
                },
                "seed": {
                    "type": "integer"
                },
                "service_tier": {
                    "description": "Specifies the latency tier to use for processing the request.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/openai.ServiceTier"
                        }
                    ]
                },
                "stop": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "store": {
                    "description": "Store controls whether the latest input and generated response should be persisted",
                    "type": "boolean"
                },
                "store_reasoning": {
                    "description": "StoreReasoning controls whether reasoning content (if present) should also be persisted",
                    "type": "boolean"
                },
                "stream": {
                    "type": "boolean"
                },
                "stream_options": {
                    "description": "Options for streaming response. Only set this when you set stream: true.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/openai.StreamOptions"
                        }
                    ]
                },
                "temperature": {
                    "type": "number"
                },
                "tool_choice": {
                    "description": "This can be either a string or an ToolChoice object."
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/openai.Tool"
                    }
                },
                "top_logprobs": {
                    "description": "TopLogProbs is an integer between 0 and 5 specifying the number of most likely tokens to return at each\ntoken position, each with an associated log probability.\nlogprobs must be set to true if this parameter is used.",
                    "type": "integer"
                },
                "top_p": {
                    "type": "number"
                },
                "user": {
                    "type": "string"
                },
                "verbosity": {
                    "description": "Verbosity determines how many output tokens are generated. Lowering the number of\ntokens reduces overall latency. It can be set to \"low\", \"medium\", or \"high\".\nNote: This field is only confirmed to work with gpt-5, gpt-5-mini and gpt-5-nano.\nAlso, it is not in the API reference of chat completion at the time of writing,\nthough it is supported by the API.",
                    "type": "string"
                }
            }
        },
        "chatrequests.ConversationReference": {
            "type": "object"
        },
...
        "modelresponses.ModelResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
...
        "modelresponses.ProviderResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "base_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "modelresponses.ProviderResponseList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelresponses.ProviderResponse"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
        "modelresponses.ProviderWithModelCountResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "base_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "model_active_count": {
                    "type": "integer"
                },
                "model_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "modelresponses.ProviderWithModelsResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "base_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelresponses.ModelResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
...
        "openai.ChatCompletionResponseFormat": {
            "type": "object",
            "properties": {
                "json_schema": {
                    "$ref": "#/definitions/openai.ChatCompletionResponseFormatJSONSchema"
                },
                "type": {
                    "$ref": "#/definitions/openai.ChatCompletionResponseFormatType"
                }
            }
        },
...
        "openai.FunctionDefinition": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "Parameters is an object describing the function.\nYou can pass json.RawMessage to describe the schema,\nor you can pass in a struct which serializes to the proper JSON schema.\nThe jsonschema package is provided for convenience, but you should\nconsider another specialized library if you require more complex schemas."
                },
                "strict": {
                    "type": "boolean"
                }
            }
        },
...
        "openai.Prediction": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "openai.ServiceTier": {
            "type": "string",
            "enum": [
                "auto",
                "default",
                "flex",
                "priority"
            ],
            "x-enum-varnames": [
                "ServiceTierAuto",
                "ServiceTierDefault",
                "ServiceTierFlex",
                "ServiceTierPriority"
            ]
        },
...
        "requestmodels.AddProviderRequest": {
            "type": "object",
            "required": [
                "base_url",
                "name",
                "vendor"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "api_key": {
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
...
        "requestmodels.UpdateProviderRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "api_key": {
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "UUID from PlatformError",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}